# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blog API",
    "description": "\n## 博客系统 API\n\n一个功能完整的现代化博客系统后端 API，支持 Markdown/MDX 内容管理和 GitOps 自动化同步。\n\n### 核心特性\n\n- 🔐 **用户认证**：基于 JWT 的安全认证系统\n- 📝 **内容管理**：支持 Markdown/MDX 格式的文章创建和编辑\n- 🎨 **三种渲染模式**：后端渲染、前端 SSR、前端 CSR\n- 📁 **媒体管理**：图片、视频等媒体文件的上传和管理\n- 🔄 **GitOps 同步**：Git 仓库与数据库的双向自动化同步\n- 🔗 **Webhook 集成**：支持 GitHub Webhook 自动触发同步\n\n### 技术栈\n\n- **框架**：FastAPI + SQLModel + PostgreSQL\n- **认证**：JWT + OAuth2\n- **内容处理**：markdown-it-py + MDX\n- **异步支持**：asyncio + aiofiles\n\n### 快速开始\n\n1. 注册账号：`POST /api/v1/users/register`\n2. 登录获取 token：`POST /api/v1/users/login`\n3. 使用 token 访问受保护的接口：`Authorization: Bearer <token>`\n\n### 文档\n\n- [Swagger UI](/docs) - 交互式 API 文档\n- [Scalar UI](/scalar) - 现代化 API 文档\n- [OpenAPI JSON](/openapi.json) - OpenAPI 规范文件\n    ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/posts/preview": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "文章实时预览",
        "operationId": "previewPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post_type}": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "创建文章",
        "operationId": "createPostByType",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostType",
              "description": "板块类型"
            },
            "description": "板块类型"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "获取指定板块的文章列表",
        "description": "获取指定板块的文章列表（自动分页）\n\n公开接口，只显示已发布的文章。\n\n示例：\n- GET /posts/article?page=1&size=20 - 文章列表\n- GET /posts/idea?page=1&size=20 - 想法列表\n- GET /posts/article?category_id=xxx - 按分类筛选",
        "operationId": "listPostsByType",
        "parameters": [
          {
            "name": "post_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostType",
              "description": "文章类型"
            },
            "description": "文章类型"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "文章状态（draft/published/archived）",
              "title": "Status"
            },
            "description": "文章状态（draft/published/archived）"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类ID",
              "title": "Category Id"
            },
            "description": "分类ID"
          },
          {
            "name": "tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签ID",
              "title": "Tag Id"
            },
            "description": "标签ID"
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "作者ID",
              "title": "Author Id"
            },
            "description": "作者ID"
          },
          {
            "name": "is_featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否为推荐文章",
              "title": "Is Featured"
            },
            "description": "是否为推荐文章"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索关键词（标题、内容）",
              "title": "Search"
            },
            "description": "搜索关键词（标题、内容）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PostShortResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post_type}/{post_id}": {
      "patch": {
        "tags": [
          "posts"
        ],
        "summary": "更新文章",
        "operationId": "updatePostByType",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostType",
              "description": "板块类型"
            },
            "description": "板块类型"
          },
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "删除文章",
        "operationId": "deletePostByType",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostType",
              "description": "板块类型"
            },
            "description": "板块类型"
          },
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "通过ID获取文章详情",
        "description": "根据 UUID 获取文章详情并增加浏览量\n\n权限规则：\n- 已发布文章：任何人可访问（包括未登录）\n- 草稿文章：只有作者或超级管理员可访问\n\n查询参数：\n- include_mdx: 是否包含原始 MDX 内容（用于编辑页面）\n  - False（默认）: 返回 AST（节省带宽）\n  - True: 返回 MDX（用于编辑）\n\n示例：\n- GET /posts/article/{uuid}\n- GET /posts/article/{uuid}?include_mdx=true\n\n注意：使用 :uuid 路径转换器确保与 slug 路由不冲突",
        "operationId": "getPostById",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostType",
              "description": "板块类型"
            },
            "description": "板块类型"
          },
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          },
          {
            "name": "include_mdx",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "是否包含原始 MDX 内容（用于编辑）",
              "default": false,
              "title": "Include Mdx"
            },
            "description": "是否包含原始 MDX 内容（用于编辑）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "timestamp",
          "request_id"
        ],
        "title": "ErrorDetail",
        "description": "错误详情模型"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "统一错误响应模型"
      },
...
      "Page_PostShortResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostShortResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[PostShortResponse]"
      },
...
      "PostCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "post_type": {
            "$ref": "#/components/schemas/PostType",
            "default": "article"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus",
            "default": "draft"
          },
          "content_mdx": {
            "type": "string",
            "title": "Content Mdx",
            "description": "原始 MDX 内容"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "cover_media_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Media Id"
          },
          "is_featured": {
            "type": "boolean",
            "title": "Is Featured",
            "default": false
          },
          "allow_comments": {
            "type": "boolean",
            "title": "Allow Comments",
            "default": true
          },
          "enable_jsx": {
            "type": "boolean",
            "title": "Enable Jsx",
            "default": false
          },
          "use_server_rendering": {
            "type": "boolean",
            "title": "Use Server Rendering",
            "default": true
          },
          "meta_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Title"
          },
          "meta_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Description"
          },
          "meta_keywords": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Keywords"
          },
          "git_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Hash"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "commit_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Message"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content_mdx"
        ],
        "title": "PostCreate"
      },
...
      "PostPreviewRequest": {
        "properties": {
          "content_mdx": {
            "type": "string",
            "title": "Content Mdx",
            "description": "要预览的 MDX 内容"
          }
        },
        "type": "object",
        "required": [
          "content_mdx"
        ],
        "title": "PostPreviewRequest",
        "description": "文章预览请求"
      },
      "PostPreviewResponse": {
        "properties": {
          "content_ast": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content Ast"
          },
          "toc": {
            "items": {},
            "type": "array",
            "title": "Toc"
          },
          "reading_time": {
            "type": "integer",
            "title": "Reading Time"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          }
        },
        "type": "object",
        "required": [
          "content_ast",
          "toc",
          "reading_time",
          "excerpt"
        ],
        "title": "PostPreviewResponse",
        "description": "文章预览响应"
      },
...
      "PostStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ],
        "title": "PostStatus",
        "description": "文章状态枚举"
      },
      "PostType": {
        "type": "string",
        "enum": [
          "article",
          "idea"
        ],
        "title": "PostType",
        "description": "内容类型枚举"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/users/login"
          }
        }
      }
    }
  },
...
}